util: move to app directory
