+++ app/generate-password.go
@@ -1,4 +1,4 @@
-package util
+package app
 
 import (
 	"net/http"
