+++ internal/cron/cron.go
@@ -1,7 +1,7 @@
 package cron
 
 import (
-	"github.com/pass-wall/passwall-server/util"
+	"github.com/pass-wall/passwall-server/app"
 	"github.com/robfig/cron/v3"
 )
 
@@ -9,6 +9,6 @@ import (
 func Setup() {
 	cron := cron.New()
 	// TODO: This 24h option should be on config file with hours format.
-	cron.AddFunc("@every 24h", func() { util.BackupData() })
+	cron.AddFunc("@every 24h", func() { app.BackupData() })
 	cron.Start()
 }
